+++ Changelog.md
@@ -17,6 +17,11 @@
 
 * Added shortcuts support for the Modify tab https://github.com/jeremytammik/RevitLookup/issues/150
 * Added EvaluatedParameter support
+* Added Category.get_Visible support
+* Added Category.GetLineWeight support
+* Added Category.GetLinePatternId support
+* Added Category.GetElements extension
+* Added Reference.ConvertToStableRepresentation support
 
 ## Bugs
 
@@ -93,7 +98,7 @@ In this release, the entire code base has been completely rewritten from scratch
 * Generic names support
 
   | Before                                                                                                                  | Now                                                                                                                   |
-      |-----------------------------------------------------------------------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------|
+        |-----------------------------------------------------------------------------------------------------------------------|-----------------------------------------------------------------------------------------------------------------------|
   | ![image](https://user-images.githubusercontent.com/20504884/225105646-37f2b052-f3fc-4771-967b-0578a94f9b07.png) | ![image](https://user-images.githubusercontent.com/20504884/225852403-4023c704-1932-471e-9f9f-84f8433013d7.png) |
 
 * Multiple results for methods with overloads